Cache icons on the server for quicker access (and also cache them in the browser)

+++ api/app/upload.js
@@ -50,7 +50,10 @@ function uploadIcon(url, share, pkg, filepath, callback) {
                     }
                     else {
                         uploadToSmartfile(url, share, filepath, iconname, function(err, url) {
-                            fs.unlink(filepath);
+                            if (iconname.indexOf('404.png') > -1) {
+                                fs.unlink(filepath);
+                            }
+
                             callback(err, url);
                         });
                     }
